rewrote sql-skript / changed pom

+++ src/main/java/com/example/springhelloworld/service/impl/GenreServiceImpl.java
@@ -26,7 +26,7 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public GenresBase update(GenresBase genre) {
-        GenresBase updateGenre = genreRepository.findByGenreId((long) genre.getGenreId());
+        GenresBase updateGenre = genreRepository.findByGenreId(genre.getGenreId());
         updateGenre.setGenreName(genre.getGenreName());
 
         return genreRepository.save(updateGenre);
